fix tests and missing wrappers required for isoaltedvm

+++ libs/util/testing/Fixture.js
@@ -13,6 +13,7 @@ const conf = {
   databaseURL: 'mongodb://localhost:27017',
   databaseName: 'testssc',
   streamNodes: ['https://api.hive.blog'],
+  defaultLogLevel: "info",
 };
 
 class Fixture {
@@ -89,8 +90,10 @@ class Fixture {
   }
 
   unloadPlugin(plugin) {
-    this.plugins[plugin.PLUGIN_NAME].cp.kill('SIGINT');
-    this.plugins[plugin.PLUGIN_NAME] = null;
+    if (this.plugins[plugin.PLUGIN_NAME]) {
+      this.plugins[plugin.PLUGIN_NAME].cp.kill('SIGINT');
+      this.plugins[plugin.PLUGIN_NAME] = null;
+    }
   }
 
   getNextTxId() {
